pyproject.toml
[project]
name = "search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.26.0",
    "google-genai>=1.66.0",
]
